Refactored folder structure to src/
