+++ .keys
@@ -1,0 +1,1 @@
+{"pubKey":"3228e785600b422f537629412983809dba83849eb2634412ee6e3feb2681ba0d","privKey":"04030000000000060000000109050909040006040007000700060300000800093228e785600b422f537629412983809dba83849eb2634412ee6e3feb2681ba0d"}
