--- database/level/blockchain/LOG
@@ -1,3 +1,0 @@
-2019/02/28-09:15:46.110595 7f8653fff700 Recovering log #23
-2019/02/28-09:15:46.128044 7f8653fff700 Delete type=0 #23
-2019/02/28-09:15:46.128089 7f8653fff700 Delete type=3 #22
