--- database/level/blockchain/LOG.old
@@ -1,3 +1,0 @@
-2019/02/28-09:10:12.809297 7f300d6fb700 Recovering log #21
-2019/02/28-09:10:12.825226 7f300d6fb700 Delete type=0 #21
-2019/02/28-09:10:12.825288 7f300d6fb700 Delete type=3 #20
