--- database/level/globalState/CURRENT
@@ -1,1 +1,0 @@
-MANIFEST-000024
