--- database/level/globalState/LOG
@@ -1,3 +1,0 @@
-2019/02/28-09:15:46.110679 7f86537fe700 Recovering log #23
-2019/02/28-09:15:46.128044 7f86537fe700 Delete type=0 #23
-2019/02/28-09:15:46.128087 7f86537fe700 Delete type=3 #22
