--- database/level/globalState/LOG.old
@@ -1,3 +1,0 @@
-2019/02/28-09:10:12.809354 7f300defc700 Recovering log #21
-2019/02/28-09:10:12.828303 7f300defc700 Delete type=0 #21
-2019/02/28-09:10:12.828342 7f300defc700 Delete type=3 #20
