+++ src/lib/p2p/index.js
