+++ src/lib/p2p/initialconnector.js
@@ -14,7 +14,7 @@ module.exports = class InitialConnector {
 
   constructor(firstTimeout) {
     if (!instance) {
-      this.ip = 'http://p2p:8082';
+      this.ip = 'http://145.93.57.202:8082';
       this.myIp = undefined;
       this.peerIp = undefined;
       this.timeout = firstTimeout;
