+++ src/lib/p2p/peer.js
