+++ src/lib/security/blockSecurity.js
