+++ src/lib/security/transactionSecurity.js
