+++ src/lib/util/keyStorage.js
