+++ src/lib/validator.js
