+++ src/services/qrCodeGenerator.js
